[CTP-OFFLOAD] Avoiding the call to Driver during suspend
BZ: 97715

Do not allow the power manager(PM) to force suspend when offload
stream is active. One of the way to avoid is to hold a wakelock
till the stream is completely finished.

In offload, at the end of stream, we wait for some time (3 sec)
before closing the stream. During that time the AF would have
released the wakelock, which will force the drv/fw to suspend
by PM. When offload try to close the stream after wait time
(3 sec),has issues of SRAM access. We are taking a wake
lock in offload HAL when stream is opened and released only
when the stream closed. Till the close is called the PM
will not force suspend the drv/fw and any call to drv/fw
will work properly

Change-Id: I6f10ced5dfb1615397a35eb086f0abe04b4d7ef8
Signed-off-by: samaga Prasanna Krishna <samaga.krishna@intel.com>
Reviewed-on: http://android.intel.com:8080/99775
Reviewed-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>
Reviewed-by: Kandasamy, Muthukumar <muthukumar.kandasamy@intel.com>
Reviewed-by: Anand, Jerome <jerome.anand@intel.com>
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>