changes made to Offload HAL for porting to KK
BZ: 155056 158018 157476 157426

The changes of all the four patches are combined and
pushed as part of this patch. The detail modifications are
155056-The sound card number is obtained only during the
opening of stream and stored in stream descriptor. The same
is used to set and get volume functions.
158018-The condition check after compress_write is modified
to handle the long playback
157476-Buffer size calculation is added
157426-AAC parameters for offload are set

Change-Id: Ifc60a162c1e488d524da1ca3d25f5f5cac2a9f5c
Category: feature differentiation
Domain: AUDIO_MEDIA-LPA
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: samaga Prasanna Krishna <samaga.krishna@intel.com>