Moving the setParameter from APM to Offload HAL
BZ: 170457

LPE pipeline is configured using a setParameter to primary
HAL. This is done from APM when ever the device is opened.
For some applications, when device is no more used, to
save power the offload HAL will closed the device during
standby. From MW point of view the device is still active.
When application resumes the playback, APM will not call
setParameter, which will result in non configuration of
LPE pipeline and no audio output in devie. To avoid this
the setParameter to primary HAL is moved from APM to
offload HAL.
This patch also changes the C implimentation of offload
HAL to C++ implimentation. This is needed to make use of
setParameter function of MW, which is in C++

Changes in this patch are
    1. Moving the setProperty from APM to offload hal ( This file)
    2. Moving the HAL code from C to C++
    3. Fixing some compilation warnings in the code

Change-Id: Iff559d8de84a6ce3b7f54318673eb67ab292acb2
Category: feature differentiation
Domain: AUDIO_MEDIA-LPA
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: samaga Prasanna Krishna <samaga.krishna@intel.com>